Updated to 0.0.3
Added a new 'Point' class to the stock package, as well as a 'Stock' class, which contains all the information for a given asset / stock. The stock class uses instances of the Point class to represent each change in the stock market. For now, we assume that points are stored date-wise.