Updated to version 0.0.5
Added the ability to simulate the trade of options stocks, and a new feature allowing users to enter a stop loss or take profit for any given position. There is an issue with data redundancies which will need to be resolved, however this will be updated in a later version.